requirements.txt
# Flask Web Framework
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1

# TensorFlow for AI Model
tensorflow==2.15.0
# Note: For production deployment on limited resources, consider tensorflow-cpu

# Image Processing
Pillow==10.1.0
numpy==1.24.3

# Firebase Admin SDK
firebase-admin==6.3.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Production Server (REQUIRED for deployment)
gunicorn==21.2.0

# Optional: Windows alternative to gunicorn
# waitress==2.1.2
